+++ components-api/src/main/scala/pl/touk/nussknacker/engine/ModelConfig.scala
@@ -77,7 +77,9 @@ object ModelConfig {
     case object SingleJsonTemplateParameter extends JsonLikeValuesEnteringMode
   }
 
-  final case class GlobalParametersConfig(editorsForStringType: NonEmptyList[ParameterEditor])
+  final case class GlobalParametersConfig(
+      editorsForStringType: NonEmptyList[ParameterEditor],
+  )
 
   object GlobalParametersConfig {
 
@@ -112,7 +114,7 @@ object ModelConfig {
       config.getAs[NonEmptyList[ParameterEditor]]("globalParametersConfig.editorsForStringType")
 
     GlobalParametersConfig(
-      editorsForStringType = maybeStringEditors.getOrElse(GlobalParametersConfig.default.editorsForStringType)
+      editorsForStringType = maybeStringEditors.getOrElse(GlobalParametersConfig.default.editorsForStringType),
     )
   }