+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/definition/ParameterValidator.scala
@@ -2,8 +2,9 @@ package pl.touk.nussknacker.engine.api.definition
 
 import cats.data.Validated
 import cats.data.Validated.{invalid, valid}
-import io.circe.generic.extras.ConfiguredJsonCodec
+import io.circe.{Codec, Decoder, DecodingFailure, Encoder, Json}
 import io.circe.parser._
+import io.circe.syntax._
 import pl.touk.nussknacker.engine.api.CirceUtil._
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
@@ -17,11 +18,21 @@ import java.util.regex.Pattern
 import scala.collection.concurrent.TrieMap
 import scala.util.Try
 
-trait Validator {
+sealed trait Validator
+
+trait CompileTimeValidator extends Validator {
 
   def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
   ): Validated[PartSubGraphCompilationError, Unit]
+
+}
+
+trait RuntimeValidator extends Validator {
+
+  def isValid(paramName: ParameterName, expression: Expression, value: Any)(
+      implicit nodeId: NodeId
+  ): Validated[Throwable, Unit]
 
 }
 
@@ -35,11 +46,56 @@ trait Validator {
   * TODO: This being sealed also makes the tests of cases that use these validators dependant on the code here -
   * not good/unseal!!
   */
-@ConfiguredJsonCodec sealed trait ParameterValidator extends Validator
+sealed trait ParameterValidator extends Validator
+
+object ParameterValidator {
+  private val fixedValuesCodec: Codec[FixedValuesValidator] =
+    Codec.forProduct1("possibleValues")(FixedValuesValidator.apply)(_.possibleValues)
+
+  private val regExpCodec: Codec[RegExpParameterValidator] =
+    Codec.forProduct3("pattern", "message", "description")(RegExpParameterValidator.apply)(v =>
+      (v.pattern, v.message, v.description)
+    )
+
+  private val minimalNumberCodec: Codec[MinimalNumberValidator] =
+    Codec.forProduct1("minimalNumber")(MinimalNumberValidator.apply)(_.minimalNumber)
+
+  private val maximalNumberCodec: Codec[MaximalNumberValidator] =
+    Codec.forProduct1("maximalNumber")(MaximalNumberValidator.apply)(_.maximalNumber)
+
+  private val validationExpressionCodec: Codec[ValidationExpressionParameterValidatorToCompile] =
+    Codec.forProduct2("validationExpression", "validationFailedMessage")(
+      ValidationExpressionParameterValidatorToCompile.apply
+    )(v => (v.validationExpression, v.validationFailedMessage))
+
+  implicit def decoder: Decoder[ParameterValidator] = Decoder.instance { cursor =>
+    cursor.downField("type").as[String].flatMap {
+      case "MandatoryParameterValidator"        => Right(MandatoryParameterValidator)
+      case "NotNullParameterValidator"          => Right(NotNullParameterValidator)
+      case "CompileTimeEvaluableValueValidator" => Right(CompileTimeEvaluableValueValidator)
+      case "NotBlankParameterValidator"         => Right(NotBlankParameterValidator)
+      case "LiteralIntegerValidator"            => Right(LiteralIntegerValidator)
+      case "JsonValidator"                      => Right(JsonValidator)
+      case "FixedValuesValidator"               => fixedValuesCodec(cursor)
+      case "RegExpParameterValidator"           => regExpCodec(cursor)
+      case "MinimalNumberValidator"             => minimalNumberCodec(cursor)
+      case "MaximalNumberValidator"             => maximalNumberCodec(cursor)
+      case "ValidationExpressionParameterValidatorToCompile" =>
+        validationExpressionCodec(cursor)
+      case other =>
+        Left(DecodingFailure(s"Unknown ParameterValidator type: $other", cursor.history))
+    }
+  }
+
+}
+
+trait CompileTimeParameterValidator extends ParameterValidator with CompileTimeValidator
+
+trait RuntimeParameterValidator extends ParameterValidator with RuntimeValidator
 
 //TODO: These validators should be moved to separated module
 
-case object MandatoryParameterValidator extends ParameterValidator {
+case object MandatoryParameterValidator extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
@@ -63,7 +119,7 @@ case object MandatoryParameterValidator extends ParameterValidator {
 
 }
 
-case object NotNullParameterValidator extends ParameterValidator {
+case object NotNullParameterValidator extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
@@ -83,7 +139,7 @@ case object NotNullParameterValidator extends ParameterValidator {
 
 }
 
-case object CompileTimeEvaluableValueValidator extends ParameterValidator {
+case object CompileTimeEvaluableValueValidator extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
@@ -104,7 +160,7 @@ case object CompileTimeEvaluableValueValidator extends ParameterValidator {
 
 }
 
-case object NotBlankParameterValidator extends ParameterValidator {
+case object NotBlankParameterValidator extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
@@ -125,7 +181,7 @@ case object NotBlankParameterValidator extends ParameterValidator {
 
 }
 
-case class FixedValuesValidator(possibleValues: List[FixedExpressionValue]) extends ParameterValidator {
+case class FixedValuesValidator(possibleValues: List[FixedExpressionValue]) extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
@@ -143,7 +199,8 @@ case class FixedValuesValidator(possibleValues: List[FixedExpressionValue]) exte
 
 }
 
-case class RegExpParameterValidator(pattern: String, message: String, description: String) extends ParameterValidator {
+case class RegExpParameterValidator(pattern: String, message: String, description: String)
+    extends CompileTimeParameterValidator {
 
   lazy val regexpPattern: Pattern = Pattern.compile(pattern)
 
@@ -163,7 +220,7 @@ case class RegExpParameterValidator(pattern: String, message: String, descriptio
 
 // TODO: we need this validator because scenario properties do not have typing result, so we enforce proper type
 //   here in validator by parsing raw expression to int
-case object LiteralIntegerValidator extends ParameterValidator {
+case object LiteralIntegerValidator extends CompileTimeParameterValidator {
 
   // empty expression should not be validated - we want to chain validators
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
@@ -185,7 +242,7 @@ case object LiteralIntegerValidator extends ParameterValidator {
 
 }
 
-case class MinimalNumberValidator(minimalNumber: BigDecimal) extends ParameterValidator {
+case class MinimalNumberValidator(minimalNumber: BigDecimal) extends CompileTimeParameterValidator {
 
   // null value should not be validated - we want to chain validators
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
@@ -208,7 +265,7 @@ case class MinimalNumberValidator(minimalNumber: BigDecimal) extends ParameterVa
 
 }
 
-case class MaximalNumberValidator(maximalNumber: BigDecimal) extends ParameterValidator {
+case class MaximalNumberValidator(maximalNumber: BigDecimal) extends CompileTimeParameterValidator {
 
   // null value should not be validated - we want to chain validators
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
@@ -234,7 +291,7 @@ case class MaximalNumberValidator(maximalNumber: BigDecimal) extends ParameterVa
 
 // This validator is not determined by default in components based on usage of JsonParameterEditor because someone may want to use only
 // editor for syntax highlight but don't want to use validator e.g. when want user to provide SpEL literal map
-case object JsonValidator extends ParameterValidator {
+case object JsonValidator extends CompileTimeParameterValidator {
 
   // null value should not be validated - we want to chain validators
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
@@ -268,7 +325,7 @@ case object JsonValidator extends ParameterValidator {
 case class ValidationExpressionParameterValidatorToCompile(
     validationExpression: Expression,
     validationFailedMessage: Option[String]
-) extends ParameterValidator {
+) extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: NodeId
@@ -290,20 +347,26 @@ object ValidationExpressionParameterValidatorToCompile {
 
 }
 
-trait CustomParameterValidator extends Validator {
+sealed trait CustomParameterValidator extends ParameterValidator {
   def name: String
 }
 
-case class CustomParameterValidatorDelegate(name: String) extends ParameterValidator {
-  import CustomParameterValidatorDelegate._
+trait CustomCompileTimeParameterValidator extends CustomParameterValidator with CompileTimeParameterValidator
 
-  override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
-      implicit nodeId: NodeId
-  ): Validated[PartSubGraphCompilationError, Unit] = getOrLoad(name).isValid(paramName, expression, value, label)
+trait CustomRuntimeParameterValidator extends CustomParameterValidator with RuntimeParameterValidator
+
+sealed trait CustomParameterValidatorLoader {
+  def load(): CustomParameterValidator
+}
+
+case class CustomParameterValidatorByNameLoader(name: String) extends CustomParameterValidatorLoader {
+  import CustomParameterValidatorByNameLoader._
+
+  override def load(): CustomParameterValidator = getOrLoad(name)
 
 }
 
-object CustomParameterValidatorDelegate {
+object CustomParameterValidatorByNameLoader {
   import scala.jdk.CollectionConverters._
 
   private val cache: TrieMap[String, CustomParameterValidator] = TrieMap[String, CustomParameterValidator]()
@@ -320,5 +383,34 @@ object CustomParameterValidatorDelegate {
     case Nil      => throw new RuntimeException(s"Cannot load custom validator: $name")
     case _        => throw new RuntimeException(s"Multiple custom validators with name: $name")
   }
+
+}
+
+case class CustomParameterValidatorByClassLoader(validatorClassName: String) extends CustomParameterValidatorLoader {
+  import CustomParameterValidatorByClassLoader._
+
+  override def load(): CustomParameterValidator = getOrLoad(validatorClassName)
+
+}
+
+object CustomParameterValidatorByClassLoader {
+
+  def apply(clazz: Class[_ <: CustomParameterValidator]): CustomParameterValidatorByClassLoader =
+    new CustomParameterValidatorByClassLoader(clazz.getName)
+
+  private val cache: TrieMap[String, CustomParameterValidator] = TrieMap[String, CustomParameterValidator]()
+
+  private def getOrLoad(className: String): CustomParameterValidator = cache.getOrElseUpdate(className, load(className))
+
+  private def load(className: String): CustomParameterValidator =
+    try {
+      Class.forName(className).getDeclaredConstructor().newInstance() match {
+        case v: CustomParameterValidator => v
+        case _ => throw new RuntimeException(s"Class $className does not extend CustomParameterValidator")
+      }
+    } catch {
+      case e: ReflectiveOperationException =>
+        throw new RuntimeException(s"Failed to instantiate CustomParameterValidator '$className': ${e.getMessage}", e)
+    }
 
 }