+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/periodic/cron/CronParameterValidator.scala
@@ -5,13 +5,13 @@ import cats.data.Validated.{invalid, valid}
 import pl.touk.nussknacker.engine.api
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.CustomParameterValidationError
-import pl.touk.nussknacker.engine.api.definition.CustomParameterValidatorDelegate
+import pl.touk.nussknacker.engine.api.definition.{CompileTimeParameterValidator, CompileTimeValidator}
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.ui.process.periodic.utils.SchedulePropertyExtractorUtils
 
 // Valid expression is e.g.: 0 * * * * ? * which means run every minute at 0 second
-object CronParameterValidator extends CustomParameterValidatorDelegate("cron_validator") {
+object CronParameterValidator extends CompileTimeParameterValidator {
 
   override def isValid(paramName: ParameterName, expression: Expression, value: Option[Any], label: Option[String])(
       implicit nodeId: api.NodeId