+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/validation/ScenarioPropertiesValidator.scala
@@ -6,7 +6,7 @@ import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.component.ScenarioPropertyConfig
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.MissingRequiredProperty
-import pl.touk.nussknacker.engine.api.definition.{MandatoryParameterValidator, ParameterValidator}
+import pl.touk.nussknacker.engine.api.definition.{CompileTimeValidator, MandatoryParameterValidator, ParameterValidator}
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.restmodel.validation.PrettyValidationErrors
@@ -57,7 +57,7 @@ class ScenarioPropertiesValidator(
     } yield (property, config.get(property._1), validator)
 
     propertiesWithConfiguredValidator
-      .collect { case (property, Some(config), validator: ParameterValidator) =>
+      .collect { case (property, Some(config), validator: CompileTimeValidator) =>
         val expression = property._2
         val value = expression match {
           case ex if ex.isBlank => null