+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/NodeCompiler.scala
@@ -126,7 +126,12 @@ class NodeCompiler(
     new StaticComponentOutputValidationContextDeterminer(globalVariablesPreparer)
 
   private val parametersEvaluator =
-    new ParameterEvaluator(globalVariablesPreparer, listeners)
+    new ParameterEvaluator(
+      globalVariablesPreparer,
+      listeners,
+      expressionCompiler,
+    )
+
   private val factory = new ComponentExecutorFactory(parametersEvaluator)
 
   private val dynamicNodeValidator =